// src/api/models/ticketHistoryItemResDTO.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { TicketHistoryItemResDTOExpiredAt } from './ticketHistoryItemResDTOExpiredAt';
import type { TicketHistoryItemResDTOSource } from './ticketHistoryItemResDTOSource';
import type { TicketHistoryItemResDTOStatus } from './ticketHistoryItemResDTOStatus';
import type { TicketHistoryItemResDTOType } from './ticketHistoryItemResDTOType';
import type { TicketHistoryItemResDTOUsedAt } from './ticketHistoryItemResDTOUsedAt';

export interface TicketHistoryItemResDTO {
  ticketId: number;
  type: TicketHistoryItemResDTOType;
  status: TicketHistoryItemResDTOStatus;
  source: TicketHistoryItemResDTOSource;
  createdAt: string;
  /** @nullable */
  usedAt?: TicketHistoryItemResDTOUsedAt;
  /** @nullable */
  expiredAt?: TicketHistoryItemResDTOExpiredAt;
}