// src/api/models/ticketHistoryItemResDTOStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

export type TicketHistoryItemResDTOStatus = typeof TicketHistoryItemResDTOStatus[keyof typeof TicketHistoryItemResDTOStatus];


export const TicketHistoryItemResDTOStatus = {
  AVAILABLE: 'AVAILABLE',
  USED: 'USED',
  EXPIRED: 'EXPIRED',
} as const;