// src/api/models/ticketHistoryItemResDTOType.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

export type TicketHistoryItemResDTOType = typeof TicketHistoryItemResDTOType[keyof typeof TicketHistoryItemResDTOType];


export const TicketHistoryItemResDTOType = {
  EXPERIENCE: 'EXPERIENCE',
  PORTFOLIO_CORRECTION: 'PORTFOLIO_CORRECTION',
} as const;